fix: modify export path and rename cli program file (#4)
