+++ bin/index.js
@@ -1,3 +1,3 @@
 #!/usr/bin/env node
-import { run } from '../dist/index.mjs';
+import { run } from '../dist/cli.mjs';
 run();
